script: Refactoring of `supportedAlgorithms` of WebCrypto
WebCrypto API has an internal object `supportedAlgorithms`
<https://w3c.github.io/webcrypto/#dfn-supportedAlgorithms> that lists
the name of all supported algorithms, with their supported operations
and desired IDL dictionary types.

We currently implement it as a long list of Rust pattern matching on the
algorithm name in string form, in the `normalize_algorithm` function.
This patch properly implements the internal object `supportedAlgorithms`
as an enum `SupportedAlgorithm`.

By doing so, we can replace a lot of string comparison, with enum
matching, which is more efficient. The long list of pattern matching is
also re-structured as a nested `match` expression so that it can be
resolved faster, compared with running through the entire list.
Furthermore, the `exportKey()` method can now properly check against the
algorithm registration, instead of hard-coding the available algorithms
separately.

Signed-off-by: Kingsley Yung <kingsley@kkoyung.dev>