ohos CI: Improve visibility of failed steps (#41551)
Using `continue-on-error: true` for a specfic step in a job, has the
huge downside that the step is marked as green / sucess in the UI, which
can make it very confusing for people trying to figure out why / where
the job failed.
This PR avoids this by removing `continue-on-error` from steps in the
bench job. Instead, we move all setup code to a place before we run the
benchmarks and add appropriate `if` conditions to later jobs, so
that we can still run both benchmarks, even if the first one fails.


Testing: 
- [mach try
ohos](https://github.com/servo/servo/actions/runs/20558990407), with
passing tests
- [mach try
ohos](https://github.com/servo/servo/actions/runs/20559189961) with
force-failed speedometer step

---------

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>